refactor LogFileNamePathParameter and update TaskManagerCustomLogHandler,JobManagerLogListHandler
