+++ flink-runtime/src/main/java/org/apache/flink/runtime/rest/messages/cluster/FileMessageParameters.java
@@ -19,10 +19,10 @@
 package org.apache.flink.runtime.rest.messages.cluster;
 
 import org.apache.flink.runtime.rest.handler.cluster.JobManagerCustomLogHandler;
+import org.apache.flink.runtime.rest.messages.LogFileNamePathParameter;
 import org.apache.flink.runtime.rest.messages.MessageParameters;
 import org.apache.flink.runtime.rest.messages.MessagePathParameter;
 import org.apache.flink.runtime.rest.messages.MessageQueryParameter;
-import org.apache.flink.runtime.rest.messages.taskmanager.LogFileNamePathParameter;
 
 import java.util.Arrays;
 import java.util.Collection;
